chore: Code splitting and refactoring of the Tx Detail and BumpFee modal
