+++ packages/suite/src/views/wallet/send/Options/BitcoinOptions/CoinControl/UtxoSelectionList/UtxoSelection/UtxoSelection.tsx
@@ -203,7 +203,7 @@ export const UtxoSelection = ({ transaction, utxo }: UtxoSelectionProps) => {
     const showTransactionDetail: MouseEventHandler = e => {
         e.stopPropagation(); // do not trigger the checkbox
         if (transaction) {
-            dispatch(openModal({ type: 'transaction-detail', tx: transaction }));
+            dispatch(openModal({ type: 'transaction-detail', tx: transaction, flow: 'detail' }));
         }
     };
